requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.12.4
importlib-metadata==4.8.1
Markdown==3.3.4
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.6.0
